Fixed and refactored tests
