+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DoOnCompleteSuite.scala
@@ -17,6 +17,7 @@
 
 package monix.reactive.internal.operators
 
+import cats.effect.IO
 import minitest.TestSuite
 import monix.eval.Task
 import monix.execution.Ack
@@ -36,11 +37,27 @@ object DoOnCompleteSuite extends TestSuite[TestScheduler] {
       "TestScheduler should have no pending tasks")
   }
 
+  test("should work for cats.effect.IO") { implicit s =>
+    var wasTriggered = 0
+    var wasCompleted = 0
+
+    Observable.now(1).doOnCompleteF(IO { wasTriggered += 1 })
+      .unsafeSubscribeFn(new Subscriber[Int] {
+        val scheduler = s
+        def onNext(elem: Int) = Continue
+        def onError(ex: Throwable): Unit = ()
+        def onComplete(): Unit = wasCompleted += 1
+      })
+
+    assertEquals(wasTriggered, 1)
+    assertEquals(wasCompleted, 1)
+  }
+
   test("should work for synchronous subscribers") { implicit s =>
     var wasTriggered = 0
     var wasCompleted = 0
 
-    Observable.now(1).doOnCompleteF(() => wasTriggered += 1)
+    Observable.now(1).doOnComplete(Task.eval { wasTriggered += 1 })
       .unsafeSubscribeFn(new Subscriber[Int] {
         val scheduler = s
         def onNext(elem: Int) = Continue
@@ -56,7 +73,7 @@ object DoOnCompleteSuite extends TestSuite[TestScheduler] {
     var wasTriggered = 0
     var wasCompleted = 0
 
-    Observable.now(1).doOnCompleteF(() => wasTriggered += 1)
+    Observable.now(1).doOnComplete(Task.eval { wasTriggered += 1 })
       .unsafeSubscribeFn(new Subscriber[Int] {
         val scheduler = s
         def onNext(elem: Int) = Future(Continue)
@@ -75,7 +92,7 @@ object DoOnCompleteSuite extends TestSuite[TestScheduler] {
     var wasCompleted = 0
     var errorThrown: Throwable = null
 
-    Observable.raiseError(dummy).doOnCompleteF(() => wasTriggered += 1)
+    Observable.raiseError(dummy).doOnComplete(Task.eval { wasTriggered += 1 })
       .unsafeSubscribeFn(new Subscriber[Long] {
         val scheduler = s
         def onNext(elem: Long): Future[Ack] =
@@ -96,7 +113,7 @@ object DoOnCompleteSuite extends TestSuite[TestScheduler] {
   test("should be cancelable") { implicit s =>
     var wasTriggered = 0
     val cancelable = Observable.now(1).delayOnNext(1.second)
-      .doOnCompleteF(() => wasTriggered += 1)
+      .doOnComplete(Task.eval { wasTriggered += 1 })
       .subscribe()
 
     s.tick()
@@ -110,7 +127,7 @@ object DoOnCompleteSuite extends TestSuite[TestScheduler] {
     val dummy = DummyException("dummy")
     var errorThrown: Throwable = null
 
-    Observable.now(1).doOnComplete(Task.raiseError(dummy))
+    Observable.now(1).doOnComplete(Task.eval { throw dummy })
       .unsafeSubscribeFn(new Subscriber[Int] {
         val scheduler = s
