+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DoOnErrorSuite.scala
@@ -19,12 +19,14 @@ package monix.reactive.internal.operators
 
 import cats.effect.IO
 import minitest.TestSuite
+import monix.eval.Task
 import monix.execution.Ack
 import monix.execution.Ack.Continue
 import monix.execution.schedulers.TestScheduler
 import monix.reactive.Observable
 import monix.execution.exceptions.DummyException
 import monix.reactive.observers.Subscriber
+
 import scala.concurrent.Future
 import scala.concurrent.duration._
 
@@ -37,12 +39,27 @@ object DoOnErrorSuite extends TestSuite[TestScheduler] {
 
   val dummy = DummyException("ex")
 
+  test("should work for cats.effect.IO") { implicit s =>
+    var wasTriggered: Throwable = null
+    var wasCompleted = 0
+
+    Observable.now(1).endWithError(dummy).doOnErrorF(ex => IO { wasTriggered = ex })
+      .unsafeSubscribeFn(new Subscriber[Int] {
+        val scheduler = s
+        def onNext(elem: Int) = Continue
+        def onError(ex: Throwable): Unit = wasCompleted += 1
+        def onComplete(): Unit = ()
+      })
+
+    assertEquals(wasCompleted, 1)
+    assertEquals(wasTriggered, dummy)
+  }
+
   test("should work for synchronous subscribers") { implicit s =>
     var wasTriggered: Throwable = null
     var wasCompleted = 0
 
-    Observable.now(1).endWithError(dummy)
-      .doOnErrorF(ex => IO { wasTriggered = ex })
+    Observable.now(1).endWithError(dummy).doOnError(ex => Task.eval { wasTriggered = ex })
       .unsafeSubscribeFn(new Subscriber[Int] {
         val scheduler = s
         def onNext(elem: Int) = Continue
@@ -58,8 +75,7 @@ object DoOnErrorSuite extends TestSuite[TestScheduler] {
     var wasTriggered: Throwable = null
     var wasCompleted = 0
 
-    Observable.now(1).endWithError(dummy)
-      .doOnErrorF(ex => IO { wasTriggered = ex })
+    Observable.now(1).endWithError(dummy).doOnError(ex => Task.eval { wasTriggered = ex })
       .unsafeSubscribeFn(new Subscriber[Int] {
         val scheduler = s
         def onNext(elem: Int) = Future(Continue)
@@ -76,8 +92,7 @@ object DoOnErrorSuite extends TestSuite[TestScheduler] {
     var wasTriggered = 0
     var wasCompleted = 0
 
-    Observable.range(0,10)
-      .doOnErrorF(_ => IO { wasTriggered += 1 })
+    Observable.range(0,10).doOnError(_ => Task.eval { wasTriggered += 1 })
       .unsafeSubscribeFn(new Subscriber[Long] {
         val scheduler = s
         def onNext(elem: Long): Future[Ack] =
@@ -98,7 +113,7 @@ object DoOnErrorSuite extends TestSuite[TestScheduler] {
     val cancelable = Observable.now(1)
       .delayOnNext(1.second)
       .endWithError(dummy)
-      .doOnErrorF(_ => IO(wasTriggered += 1))
+      .doOnError(_ => Task.eval { wasTriggered += 1 })
       .subscribe()
 
     s.tick()
@@ -113,7 +128,7 @@ object DoOnErrorSuite extends TestSuite[TestScheduler] {
     val dummy2 = DummyException("dummy2")
     var errorThrown: Throwable = null
 
-    Observable.now(1).endWithError(dummy2).doOnError(_ => throw dummy1)
+    Observable.now(1).endWithError(dummy2).doOnError(_ => Task.eval { throw dummy1 })
       .unsafeSubscribeFn(new Subscriber[Int] {
         val scheduler = s
