+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DoOnNextSuite.scala
@@ -17,6 +17,7 @@
 
 package monix.reactive.internal.operators
 
+import cats.effect.IO
 import minitest.TestSuite
 import monix.eval.Task
 import monix.execution.Ack.Continue
@@ -32,11 +33,45 @@ object DoOnNextSuite extends TestSuite[TestScheduler] {
       "TestScheduler should have no pending tasks")
   }
 
-  test("should work for Observable.range") { implicit s =>
+  test("should work for cats.effect.IO") { implicit s =>
     var sum = 0L
     var wasCompleted = 0
 
-    Observable.range(0, 20).doOnNext(x => Task { sum += x })
+    Observable.range(0, 20).doOnNextF(x => IO(sum += x))
+      .unsafeSubscribeFn(new Subscriber[Long] {
+        val scheduler = s
+        def onError(ex: Throwable): Unit = ()
+        def onNext(elem: Long) = Continue
+        def onComplete(): Unit = wasCompleted += 1
+      })
+
+    s.tick()
+    assertEquals(sum, 190)
+    assertEquals(wasCompleted, 1)
+  }
+
+  test("should work for Observable.range and async task") { implicit s =>
+    var sum = 0L
+    var wasCompleted = 0
+
+    Observable.range(0, 20).doOnNext(x => Task.evalAsync(sum += x))
+      .unsafeSubscribeFn(new Subscriber[Long] {
+        val scheduler = s
+        def onError(ex: Throwable): Unit = ()
+        def onNext(elem: Long) = Continue
+        def onComplete(): Unit = wasCompleted += 1
+      })
+
+    s.tick()
+    assertEquals(sum, 190)
+    assertEquals(wasCompleted, 1)
+  }
+
+  test("should work for Observable.range and sync task") { implicit s =>
+    var sum = 0L
+    var wasCompleted = 0
+
+    Observable.range(0, 20).doOnNext(x => Task.eval(sum += x))
       .unsafeSubscribeFn(new Subscriber[Long] {
         val scheduler = s
         def onError(ex: Throwable): Unit = ()
@@ -48,11 +83,28 @@ object DoOnNextSuite extends TestSuite[TestScheduler] {
     assertEquals(wasCompleted, 1)
   }
 
-  test("should work for Observable.now") { implicit s =>
+  test("should work for Observable.now and async task") { implicit s =>
     var sum = 0L
     var wasCompleted = 0
 
-    Observable.now(10L).doOnNext(x => Task { sum += x })
+    Observable.now(10L).doOnNext(x => Task.evalAsync(sum += x))
+      .unsafeSubscribeFn(new Subscriber[Long] {
+        val scheduler = s
+        def onError(ex: Throwable): Unit = ()
+        def onNext(elem: Long) = Continue
+        def onComplete(): Unit = wasCompleted += 1
+      })
+
+    s.tick()
+    assertEquals(sum, 10L)
+    assertEquals(wasCompleted, 1)
+  }
+
+  test("should work for Observable.now and sync task") { implicit s =>
+    var sum = 0L
+    var wasCompleted = 0
+
+    Observable.now(10L).doOnNext(x => Task.eval(sum += x))
       .unsafeSubscribeFn(new Subscriber[Long] {
         val scheduler = s
         def onError(ex: Throwable): Unit = ()
@@ -64,13 +116,13 @@ object DoOnNextSuite extends TestSuite[TestScheduler] {
     assertEquals(wasCompleted, 1)
   }
 
-  test("should protect against user code for Observable.now") { implicit s =>
+  test("should protect against user code for direct exceptions") { implicit s =>
     val dummy = DummyException("dummy")
     var received = 0L
     var wasCompleted = 0L
     var errorThrown: Throwable = null
 
-    Observable.now(10L).doOnNext(_ => Task { throw dummy })
+    Observable.range(1,10).doOnNext(x => throw dummy)
       .unsafeSubscribeFn(new Subscriber[Long] {
         val scheduler = s
         def onError(ex: Throwable): Unit = errorThrown = ex
@@ -83,13 +135,13 @@ object DoOnNextSuite extends TestSuite[TestScheduler] {
     assertEquals(wasCompleted, 0)
   }
 
-  test("should protect against user code for Observable.range") { implicit s =>
+  test("should protect against user code for indirect exceptions") { implicit s =>
     val dummy = DummyException("dummy")
     var received = 0L
     var wasCompleted = 0L
     var errorThrown: Throwable = null
 
-    Observable.range(1,10).doOnNext(_ => Task { throw dummy })
+    Observable.range(1,10).doOnNext(x => Task.evalAsync(throw dummy))
       .unsafeSubscribeFn(new Subscriber[Long] {
         val scheduler = s
         def onError(ex: Throwable): Unit = errorThrown = ex
@@ -97,6 +149,7 @@ object DoOnNextSuite extends TestSuite[TestScheduler] {
         def onComplete(): Unit = wasCompleted += 1
       })
 
+    s.tick()
     assertEquals(received, 0)
     assertEquals(errorThrown, dummy)
     assertEquals(wasCompleted, 0)
